+++ src/combinator/provided.rs
@@ -1,7 +1,9 @@
 mod eat;
 mod next;
 mod till;
+mod wrap;
 
 pub use eat::*;
 pub use next::*;
 pub use till::*;
+pub use wrap::*;