+++ src/combinator/provided/eat.rs
@@ -12,7 +12,7 @@ use std::marker::PhantomData;
 /// See [`eat`] for more details.
 pub trait Eat {
   /// Convert the implementor to a parser implementor.
-  fn into_parser<State, Heap>(self) -> impl Parse<State, Heap, Kind = ()>;
+  fn into_parser<State, Heap>(self) -> impl Parse<Kind = (), State = State, Heap = Heap>;
 }
 
 macro_rules! impl_eat {
@@ -40,7 +40,7 @@ macro_rules! impl_eat {
 
     impl Eat for $inner {
       #[inline]
-      fn into_parser<State, Heap>(self) -> impl Parse<State, Heap, Kind = ()> {
+      fn into_parser<State, Heap>(self) -> impl Parse<Kind = (), State=State, Heap=Heap> {
         $name::new(self)
       }
     }
@@ -51,8 +51,10 @@ impl_eat!(EatChar, char, (Copy));
 impl_eat!(EatString, String, ());
 impl_eat!(EatUsize, usize, (Copy));
 
-impl<State, Heap> Parse<State, Heap> for EatChar<State, Heap> {
+impl<State, Heap> Parse for EatChar<State, Heap> {
   type Kind = ();
+  type State = State;
+  type Heap = Heap;
 
   #[inline]
   fn parse<'text>(
@@ -66,8 +68,10 @@ impl<State, Heap> Parse<State, Heap> for EatChar<State, Heap> {
   }
 }
 
-impl<State, Heap> Parse<State, Heap> for EatString<State, Heap> {
+impl<State, Heap> Parse for EatString<State, Heap> {
   type Kind = ();
+  type State = State;
+  type Heap = Heap;
 
   #[inline]
   fn parse<'text>(
@@ -81,8 +85,10 @@ impl<State, Heap> Parse<State, Heap> for EatString<State, Heap> {
   }
 }
 
-impl<State, Heap> Parse<State, Heap> for EatUsize<State, Heap> {
+impl<State, Heap> Parse for EatUsize<State, Heap> {
   type Kind = ();
+  type State = State;
+  type Heap = Heap;
 
   #[inline]
   fn parse<'text>(
@@ -116,13 +122,15 @@ impl<'a, State, Heap> EatStr<'a, State, Heap> {
 
 impl<'a> Eat for &'a str {
   #[inline]
-  fn into_parser<State, Heap>(self) -> impl Parse<State, Heap, Kind = ()> {
+  fn into_parser<State, Heap>(self) -> impl Parse<Kind = (), State = State, Heap = Heap> {
     EatStr::new(self)
   }
 }
 
-impl<'a, State, Heap> Parse<State, Heap> for EatStr<'a, State, Heap> {
+impl<'a, State, Heap> Parse for EatStr<'a, State, Heap> {
   type Kind = ();
+  type State = State;
+  type Heap = Heap;
 
   #[inline]
   fn parse<'text>(